fix: game 내용 경로 변경
